Batch Norm Consolidation
Summary: Simplify hierarchy of batch norm ops into a single,
backend agnostic op. This commit only adds the new ops, but
does not change the decomps to actually produce these ops in
the graph. This will be done after the 2 week FC window,
and the ops used in the old stack is planned to be removed
after the 6 month BC window.

ghstack-source-id: 8f68993c46cd554380ece00d60bbc19b7fc9353c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/116092