[inductor] allow padding mm/bmm/addmm in the presence of dynamic dims
Previously, pad_mm skips cases where any input tensor has symbolic
dimension or stride. This is too constraint in practise.
This PR enables this pass to pad non-symbolic dimensions in
the presence of dynamic dims. For example, with this PR, we could
pad the K dimension (i.e. 1921) for torch.mm(A[s0, 1921], B[2048, 1921]).

ghstack-source-id: fdf51b23a9dc1d37606b378423c701338d99c164
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120073